BackendServices/AuthService/requirements.txt
flask
PyJWT==1.7.1